test(api): split branch-only storage tests for the regression-proof lane
The regression-proof lane overlays this branch's tests onto the base
revision, so a modified existing test file that imports branch-only
symbols collection-errors there (an error is not proof). Move the
atomic-append / compare-and-set / empty-body-delete tests into their own
module with lazy imports, drop the now-dead read_log/write_log helpers,
and make the tracked-todo mock_deps fixture base-tolerable.

Verified with the real scripts/ci/pytest.sh regression-proof harness
(real Mongo+Redis): all marked tests fail on base as required.
